move products rendering tests to tests/rendering/ (#28018)
